requirements.txt
# Core packages
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
pyyaml>=6.0.1
pytest>=7.4.0
setuptools>=69.0.0

# MLflow for model tracking and versioning
mlflow>=2.9.0

# Airflow for workflow orchestration
apache-airflow>=2.8.0
apache-airflow-providers-redis>=3.4.0

# Redis for data storage
redis>=5.0.0
redis-py-cluster>=2.1.3

# Data validation
great-expectations>=0.17.15
pydantic>=2.5.0

# API and serving
fastapi>=0.109.0
uvicorn>=0.25.0

# Docker integration
docker>=7.0.0
